fix: move args related directories
